Iss17 load balancing (#50)
* Load-balance progress

* Documentation, MVC, load balancing, coding style

WIP #41, #32, #17, #44
Four issues got stuck in this iss17 branch, I started working on #44 in
this branch. Due to a lot of changes almost everywhere there'd be a big
chance of multiple conflicts if everything was on their own branches. So
meh

* Finished bot-san.js jsdoc

* Sending episodes and transcoding

* missed file

* Invoke episode with new

* Fixed balancing issue

* Done for testing

Completes the following checkpoints:
- When done, emit a done message to Ray with the episode object.
- When a transcoding node is disconnected, put the episodes for that
node back to the queue.

* Fix anime model

* Add status when fay received a episode.

* Fixing bugs

* Do not requeue

* Do not requeue for real this time.

+ Includes which server is sending to discord.

* 30 minutes check

Thought I had reverted this already before comitting

* Remove obsolete code.

Closes #17